Fix bug 683213 - mirror dev failure in HA LVM can cause service failure
A new option was added to 'vgreduce', --force.  This argument is now
required.

Additionally, repair is attempted when deactivating a device as well as
when trying to activate.  This is because a shutdown of a service can
race against a device failure.  Adding the repair option to the shutdown
process adds resilience.