build: ship static ocf-tester.8 man page and drop build requires on help2man
ocf-tester.8 is the only user for help2man and both creates different problems
to be either available or built on different distro/releases.

Given the simplicity of the man page, run help2man once, add it to the tree
and drop automatic build.

This also allow to drop a build require on help2man and simplify the spec file a bit

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>
Acked-By: Florian Haas <florian.haas@linbit.com>