Low: IPaddr2: remove falsely advertised default
The "nic" parameter never had "eth0" as its real default. Instead,
when "nic" is not specified, the RA tries to find out the right
interface as per the IP, netmask, and the kernel routing table, and
gives up when failing to do so.

Remove the incorrectly advertised default from the metadata.