Medium: Raid1: support for multiple MD arrays, as specified in raidconf
If raiddev is set to "auto", the RA will manage all arrays found
in the configuration file. In case one of the arrays fails,
obviously all need to be recovered.

On raidconf changes, in particular when adding new raids, the
resource needs to be unmanaged previously, then the new raids
started by hand, then managed. Otherwise, the next monitor will
fail and all running arrays restarted.