Fix wrong return value: s/OCF_RUNNING/OCF_SUCCESS/
There is no OCF_RUNNING.
return $OCF_RUNNING looks good, and will be hard to spot when reviewing
resource agents, but will typically expand to "return" (because
OCF_RUNNING is undefined), and thus return the exit code of the previous
command.

This may or may not be the indented value 0.