Low: asterisk: initialize convenience variables after validate
We don't need to set ASTRUNDIR and ASTLOGDIR for meta-data,
usage, or validate-all.
So, just set them after validate has run, when we're sure to
be able to read the config file.

Keeps ocf-tester happy, too (which tries to run meta-data as nobody,
which is expected to complete without errors or warnings).