Medium: LVM: force dmevent monitoring for clones
Starting a clustered volume with monitoring disabled is not allowed:

http://www.redhat.com/archives/lvm-devel/2010-March/msg00289.html

Which would be fine, as activation/monitoring = 1 ships as the default
in lvm.conf. However, at least some versions of LVM seem to ignore this,
throwing an error on vgchange unless "--monitor y" is explicitly set
on the command line:

https://bugs.launchpad.net/ubuntu/+source/lvm2/+bug/833368

Thus, for cloned instances, always invoke vgchange with "--monitor y".

Thanks to Nils Meyer <nils@nm.cx> for pointing out this issue.