High: iSCSITarget: treat an empty "implementation" parameter specially
If the user did not specify an "implementation" parameter, _and_
we're unable to discern the supported implementation from installed
binaries (read: we probably don't have any implementation installed),
then the default for "implementation" is empty. In that case,
_starting_ the resource is still on a specific node is still an error,
but it's not a resource configuration problem. Hence, in this case,
exit with OCF_ERR_INSTALLED, not OCF_ERR_CONFIGURED.