High: iSCSILogicalUnit: correctly match for target IQN and backing device name (tgt)
When we're dealing with several targets, or several backing devices,
where the name of one is a substring of the other, the grep/sed magic
to figure out target IDs and LUNs would fail because of multiple matches.

Fix the expressions so we match to the end of the line ("$" anchor).