low: dhcpd: Fixed PID file path for chrooted instances.
The PID file that was being used was the full chrooted path (when views
from outside the chrooted environment), which was incorrect. This caused
the dhcpd service to fail to start properly due to a failed PID file
creation.

Wrapping the PID references into the DHCPD_ARGS variable allowed for a
simple reference to be used depending on chrooted vs non-chrooted PID
file requirements.