Med: tools/sfex_lib.c: sfex_init fails on a 64 bit big endian platform
sfex_init failed on a 64 bit big endian platform.
Message at that time is the following.
 "ERROR: Failed to allocate aligned memory"

The same problem also occurs in sfex_stat,sfex_daemon that use the sfex_lib.

This issue caused by difference of the variable type.
The ioctl() system call return sector_size as a int type,
but sector_size is defined as a unsigned long type in the sfex_lib.

This patch has made the following changes to the sfex_lib.
 - add a new int value(sec_tmp) for ioctl() argument.
 - in after ioctl(), a new value(sec_tmp) is copied to sector_size as unsigned long.