resource-agents: Fix return code in fslib.sh
Patch from John Ruemker <jruemker@redhat.com>:

When you have an fs resource that depends on an lvm resource in a service,
and that lvm resource fails to start, the subsequent attempt to stop the fs
resource will fail because the device doesn't exist (since the logical
volume was never activated).

Because we're stopping a resource that is not actually running
(the fs isn't mounted), the operation should succeed here. Instead it
fails, preventing recovery.

Resolves: rhbz#853249
Signed-off-by: Ryan McCabe <rmccabe@redhat.com>