LVM agents:  Handle RAID logical volumes.
This commit addresses rhbz824153.  Specifically, it enables the use
of RAID logical volumes in HA-LVM when using the "tagging" (i.e.
not "clvm") method.

One of the advantages of using RAID LVs is that they can handle
transient failures.  If a device goes down in the RAID LV and comes
back later, it can be reintegrated into the array by simply restarting
(or relocating) the service.  Only those portions of the disk that
have changed will be resync'ed - allowing for faster recovery times.