Low: test-find: make it work properly, support IPv6, style...
Previously, the checks for values other than exit code were not
performed correctly.

The test now works exclusively with netmask form referred
to as CIDR (cidr_netmask).  Downside is that it is no longer
capable to exhibit wrong exit code of findif binary when
cidr_netmask is provided in IPv4-like notation but not in
a proper form (i.e., at least one dot is contained, but as
a whole, it is not a valid address), however this is
a deprecated usage and the issue was fixed anyway.

A very limited test set for IPv6 (targetting findif.sh)
was added.  Options such as --, -4, -6, -46 denotes
which built-in tests to use.  Additionally, one can add
ad-hoc tests definitions by piping them into the script
(for format see ./test-findif -h).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>