Medium: IPaddr2: use send_arp instead of ipoibarping if not available
The future of ipoibarping, which was used for infiniband
interfaces, seems to be uncertain and some distributions don't
have it at all. After this commit, we fallback to our send_arp,
but also issue a warning message (only on start). The new
parameter arp_sender allows for more flexibility and may be set
to send_arp to avoid the message.