Medium: exportfs: allow multiple exports
This commit enables exportfs to export multiple directories.
The parameter names remained the same which makes for an awkward
singular vs. plural in name vs. value. The alternative of adding
another parameter for multiple directories seemed more awkward.
In case of multiple directories, the fsid becomes a start_fsid,
i.e. the fsid of those directories are fsid, fsid+1, and so on.