Disable HTTPS SSL certificate hostname checking, tidy debug output
HTTPS checks with request/receive will fail under newer versions of the
LWP module due to certificate name checking being enabled by default. This
disables the check again so that service checking by IP only still succeeds.

Additional debug output shows the Status line so that it is easier to find out
why the service is down. Debug output tidied up when SSL values are not set
(usually if there has been an error, or the host is unreachable).