Low: tomcat: Put default pid file location on temporary storage
Pid files always need to be put on storage that
is refreshed after reboot. Otherwise we run the risk
of a resource's stale pid file matching some other
process when pacemaker (or some other cluster manager) is
probing the resource for activity. This can result in
unrelated processes being killed in an effort to stop
a tomcat instance that is not actually present on the system.