High: tomcat: Avoid unnecessary force kill of tomcat on stop
If the catalina.sh script is not used to start and stop
tomcat, the stop action may return before the pid has
actually exited. This results in an unnecessary kill
on the pid to occur when it would have exited on its own
had we honored the stop timeout value.