High: nfsserver: Do not require shared info directory when cloned
The shared info directory is used for Active Passive nfs server
instances. This shared directory allows for client data to failover
and follow the server instance whenever the server is reallocated to
another node.

There is another use case where the nfs server instance does not
fail over, but only the nfs exports move. When this use case is in use
the shared info directory is not necessary. Often this use-case will
involve a cloned nfs server spread out amoung several nodes. We now
allow this agent to work as a cloned resource as long as a shared info
directory is not in use.