High: ocf-shellfuncs: Log exit reason to log facilities as well as stderr
Now that we are starting to set exit reasons strings, it
is becoming obvious that most of the exit reasons are already
logged as ERRORs.  To avoid having to both set the exit string
and log the error, the ocf_exit_reason function now logs both
the exit reason to stderr, and the exit string as a priority
ERROR message to any log facilities. Since it is possible the
logfacility may just be echoing to stderr, we avoid duplicate
prints to stderr by ignoring printing to stderr when ha_log
is called from ocf_exit_reason