High: Filesystem: when loading kernel modules wait for filesystem to initialize
When the Filesystem agent is managing a filesystem type that
is not present in /proc/filesystems, the agent attempts to
load the kernel module for that filesystem.

This patch improves on that logic by
1. verifying that modprobe worked
2. give the module a brief period of time to initialize.

Item 2 is important because there is a brief period
of time between when modprobe returns loading the gfs2
module, and when gfs2 will show up in the /proc/filesystems
list.  Without retrying the search of the /proc/filesystems
file, a gfs2 filesystem may fail to start correctly because
it will look like the filesystem isn't supported.