Middle: anything: Prevent stop failure, even if the job takes time to stop.
Currentry, in the anything_stop() method, immediately monitor the job and return a value(SUCCESS or ERR) after sending SIGKILL.
But, I'm concerned about some job may not stop immediately. For example, because of Time-consuming due to the kernel inside.
So, I change the stop method to wait forever to stop the job.
This is like the stop method of pgsql RA or apache RA.