High: mysql-common: avoid use of check_binary in common validation function.
Since the environment validation exit code needs to be interpreted
differently now for monitor operations, we need to avoid functions like
'check_binary' that exit the process immediately upon failure. Instead
we should use 'have_binary' in this situation.

This allows the mysql agent to work properly in a scenario where the entire
mysql install resides on shared storage.