Medium: Xen: Use xl list $domain return code in status check (boo#901453)
Under xl, state and status are not returned with a --long listing. Instead,
the return code from `xl list $domain` must be used to determine if a domain
is running on the node or not.

As xen-list and xenstore-ls both rely on xenstore, the additiona xenstore-ls
check can safely be removed.

For POSIX compliance, the bash-specific [[ was replaced with expr.