Fix: configure : too much usage of usage of prefix
The defaults for prefix is "/usr" and for OCF_ROOT_DIR is "/usr/lib/ocf".
If the options PREFIX and OCF_ROOT_DIR of configure are used the defaults are replaced with the value of $withval.

The if condition used is wrong and only true if OCF_ROOT_DIR=prefix
[ if test x"$withval" = xprefix; then OCF_ROOT_DIR=${prefix}; else
and the better form is also nonsense
[ if test x"$withval" = x"${prefix}"; then OCF_ROOT_DIR=${prefix}; else

The result of the if condition is always the value of $withval and so unnecessary.

No trailing slashes for OCF_RA_DIR and OCF_LIB_DIR. They are always used with literals starting with a slash.

Adding the prefix to OCF_RA_DIR and OCF_LIB_DIR is unnecessary.
They are defined as absolute paths not as relative paths.