Medium: pgsql: delete an old replication slot when creating the slot.
Initialize the replication slot (delete an old slot if exists and create a new one)
to prevent that the old slot runs out of disk space.

Without the initialization, PostgreSQL would try to use the old slot after a node
failed once and re-joined to the cluster, and continue to accumulate WAL files for
the old slot, which would eventually run out of disk space.