iSCSILogicalUnit: add check for leftover target/core entries for lio-t
Apply the same checking behavior to the lio-t target as the RA
previously already did for lio: do not report the resource as being
gracefully stopped if it still has iblock handles open.

With lio-t, we cannot specify the iblock number, but since we're using
$OCF_RESOURCE_INSTANCE as the backstore key here (which is unique
across the cluster), the more inclusive file glob is still safe.

Fixes ClusterLabs/resource-agents#640.