Extend dns recive string documentation
In case of different answers on the same DNS query,
the recive string could also defined as an regex to match
multiple adresses or PTR record names.

Our use case is an DNS based ISP loadbalancing.
In situations where one of our providers has an problem our monitoring
detects this and dns answers changes acording to the adress range
of the other provider.
So we must ensure, that the health check does not disable the virtual
service in case of such an event.