Mid : Generation and error check of the container name at the time of the globally-unique designation.(repost)
The patch converts a container name at the time of the globally-unique
designation into a form supported in docker.
When docker RA cannot generate the name at the time of globally-unique
designation automatically, the patch returns an error.

If there is the method that is better for handling of sed of the patch,
I want you to change it.

Some problems still remain for the use with the clone of docker RA.
This is a temporary correction, but this correction blocks some problem
movement.
Best Regards,
Hideo Yamauchi.