ldirectord: Do an initial check of all services in when in 'starting' status
When the daemon is starting up any services will be marked as 'down' in
'starting' mode, but then marked as 'up' in running mode. This means e-mails
will be sent out saying the services has been restored, even if
emailalertstatus is set to just 'running'.

Doing an initial check of all services when starting up means everything
is already in the correct state for the time we are actually 'running'.