Medium: IPaddr2: do not send Gratuitous ARPs in the Cluster IP configuration
send_arp.linux was sending a hardware MAC instead of a multicast MAC even if
it is specified as src_hw_addr, which would prevent to work the Cluster IP properly.

It should not be necessary to send Gratuitous ARPs in the case of Cluster IP
because the multicast MAC does not change during a failover and nothing needs
to be updated on the peers.

Only when send_arp.libnet binary is used, they will be sent to retain
the old behavior.