Make use of ocf_attribute_target in rabbitmq/redis/galera
Instead of using NODENAME directly use the newly-introduced
ocf_attribute_target function. This allows the operator to decide if an RA
running inside a bundle should use per-host properties or per-bundle
properties in a resource. This can be done by setting the meta-attribute
'container-attribute-target' to 'host' in the former case and leave the
defaults as is in the latter case.

This change has been tested in the following scenarios (for rabbit/redis
and galera):
1) A deployment without bundles and without the container-attribute-target meta attr set.
2) A deployment with the resources running in bundles without the meta-attr set
3) A deployment with the resources running in bundles with the meta-attr set to 'host'

Additionally we successfully tested restarting of each resource, banning
of each resource from a node and rebooting a cluster node hosting the
resource.

Signed-off-by: Andrew Beekhof <abeekhof@redhat.com>
Signed-off-by: Michele Baldessari <michele@acksyn.org>
Signed-off-by: Damien Ciabrini <dciabrin@redhat.com>