[build] fix make distcheck by adding fallback OS dir to find docbook.xsl
on systems where xmlcatalog returns nothing, and we are running distcheck,
that enforces datadir to be outside of normal OS paths, allow /usr/share
(default datadir) as fallback to find docbook.xsl.

we continue to prefer datadir for those systems where docbook.xsl is in
non-standard locations to avoid any build regressions.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>