VirtualDomain: ensure needed virtual storage pools and networks are up and refreshed
When starting a VM or migrating it, some of the storage pools or
networks it requires may not be started. Start those if needed.

Refreshing the storage pools may also be needed. For instance when using
an pool of dir type on an OCFS2 mount, the mount may not be done at the
time of the pool start. Libvirt would then think there is no volume in
the pool...