LVM-activate: Disable VG autoactivation in system_id access_mode
The autoactivation property of volume group will be disabled when
vg_access_mode is set to system_id, to fix HA-LVM systemid race
condition, which introduces an error "active on 2 nodes".
For HA-LVM systemid use case, the node might introduce the
unexpected outage of the cluster service. In details, the node
might suffer from some race condition and reports "active on 2 nodes"
on one node after reset the active node. This will cause the active
service to be restarted which is unexpected.